Migrate all simple components into a single LayoutExtras file
Replace default exports with modue exports
